\name{phone}
\docType{data}
\alias{phone}
\title{3D data points for an image of a phone}
\description{
This dataset consists of the 3D points representing an image of a phone. The original
data was stripped of connectivity information and centroids of the triangles were added 
to generate additional points
}
\usage{data(phone)}
\format{A data.frame with 6070 rows and 3 column}
\source{\emph{J. Comput. Chem.}, 2003, \bold{24}, 1215-1221}
\keyword{datasets}
