\name{swissroll}
\docType{data}
\alias{swissroll}
\title{Data points for the Swiss Roll function in 3 dimensions}
\description{
 This dataset comprise 1000 3D points generated using:
 \deqn{x = \phi \cos{\phi}, y = \phi \sin{\phi}, z}
}
\usage{data(swissroll)}
\format{A data.frame with 1000 rows and 3 column}
\keyword{datasets}
